Change click handler of triggers to use currentTarget
The click handler used the target property of the event. This would
cause the handler not to find the matching container in case the trigger
element has child elements which could've initiated the click event.

Using the currentTarget property ensures we're using the trigger element
itself and not one if its children. This because the click handler is
specifically added to the trigger element itself.