Add aria-hidden and aria-expanded attributes
This commit adds support for the aria-hidden and aria-expanded
attributes.

The aria-expanded attributes is set on the trigger element and its value
is updated whenever the content is shown/hidden.

The aria-hidden attribute is set in the content element and its value is
also updated whenever the content is show/hidden.

Together these two attributes will help AT to better describe the state
of the trigger and content to the visitor.