Add support for aria-controls
The trigger element will be checked for the aria-controls attribute.

When this attribute is not present the module check the content element
for an ID. If an ID is present it will use this when setting the
aria-controls attribute on the trigger. When no ID is set the module
will generate a unique ID, assign it the content element and use it as
the value of the aria-controls attribute.

When the aria-controls attribute is present the module will check if it
points to an existing ID. If the ID is not of an existing element the
module will generate unique ID, assign it the content element and use it
as the value of the aria-controls attribute.

By providing the aria-controls attribute AT can provide the visitor a
way to quickly go from to the trigger to the content.

References:
- [Progressive collapsibles](http://heydonworks.com/practical_aria_examples/#progressive-collapsibles)
- [Easy ARIA Tip #5: aria-expanded and aria-controls](https://www.marcozehe.de/2010/02/10/easy-aria-tip-5-aria-expanded-and-aria-controls/)
- [Using the aria-controls attribute](http://tink.uk/using-the-aria-controls-attribute/)