Make trigger event customisable
The trigger would always react the the click event. It turned out this
was a bit too rigid, there are situations where you might want to listen
to a change event or something else.

This commit makes the trigger event configurable through the
`options.triggerEvent` option. By default it is configured to the click
event but you can change it to fit your needs.