Refactor updating the trigger text
The trigger text was updated when the element identified as the trigger
also had the attributes containing texts for the open / close state. The
assumption here was that the trigger element is always the element
responsible for displaying the text.

It turned out this assumption was wrong. There are use cases where the
trigger element is not the element showing the text but a child element
is responsible for this.

To support this use case the module has been updated to handle this
properly. It will now check of the trigger element itself has the
attribute with the state text. When the trigger element itself doesn't
have the attribute the module will now check if there is a child element
with the attribute instead.