Update README to reflect trigger text changes
The attributes which can be used to change the trigger text when the
detail view is open or closed no longer have to be present on just the
trigger element. In version 1.4 it is okay to have these attributes on
the trigger element as well as on any descendant of the trigger element.

This change had not yet been reflected in the documentation of these
options